Clear snippet preview if none provided
When switching from a grammar that provides a snippet to a grammar that does not. The snippet preview is not cleared and still shows the snippet from the unrelated grammar and wrong syntax highlighing is applied. This PR clears the snippet preview text field in case a grammar is selected that does not provide snippets.