Fix finding end of the regexp in multi-byte string
Fix finding the end of the regexp in a multi-byte string and incorrect use of bytesCount instead of the string length in multi-byte characters. In C/C++ grammars the bug is triggered by strings/comments containing multi-byte characters and numeric constants/punctuation.
Additional comment: lineTokens.produce() takes length of the string in multi-byte characters as a second parameter and captureIndices of OnigCaptureIndex are also counted as characters (see OnigNextMatchResult.captureIndicesOfMatch()).